Update default layout (#17)
1. replace logo
2. update header
3. update footer
4. update search button hover color
5. update responsive for both index page and default layout
